// src/modules/auth/schemas/user.schema.d.ts
import { Document } from 'mongoose';
import { BaseSchema } from '../../../common/schemas/base.schema';
export type UserDocument = User & Document;
export declare class User extends BaseSchema {
    email: string;
    password: string;
    role: string;
    status: string;
}
export declare const UserSchema: import("mongoose").Schema<User, import("mongoose").Model<User, any, any, any, Document<unknown, any, User, any, {}> & User & {
    _id: import("mongoose").Types.ObjectId;
} & {
    __v: number;
}, any>, {}, {}, {}, {}, import("mongoose").DefaultSchemaOptions, User, Document<unknown, {}, import("mongoose").FlatRecord<User>, {}, import("mongoose").DefaultSchemaOptions> & import("mongoose").FlatRecord<User> & {
    _id: import("mongoose").Types.ObjectId;
} & {
    __v: number;
}>;